+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - #task_number - short task description
+- #4 - entity BankCard was created, integrated into gazon-service and tested
 - #3 - created passport entity
 
 ## [1.0.0]